requirements.txt
semver==2.13.0
pygithub==1.55
